tbsecp3: get rid of pci_alloc_consistent usage
pci_alloc_consistent is a legacy function that allocates GFP_ATOMIC memory, the
driver seemingly doesn't need this and it breaks probing on some hosts (notably
the PINE64 Quartz64 Model A).

Replace this with dma_alloc_coherent using GFP_KERNEL memory, which is what is
preferred for generally all kernel allocations.

I don't have the TV tuner card to test this with, but someone who does tested
this patch for me and confirmed it to have fixed their issue.

Signed-off-by: Nicolas Frattaroli <frattaroli.nicolas@gmail.com>